misc: Documentation, better examples, and code structure
